Fix yarn install making working directory dirty
Running yarn install should not cause the working directory to have
uncommitted changes.

1) Remove examples/expo/.yarn/install-stage.gz
This file should not be checked into git. In fact, it shouldn't exist
because there shouldn't be multiple yarn.lock files in a monorepo.

I attempted to resolve this, but ran into more issues than I have time
to fix, so instead I'll just remove this so that yarn install at least
does not cause there to be uncommitted changes in the working directory.

2) Commit the change to packages/zeego/package.json

yarn install sorts the packages in package.json, so commit these
changes.