Add native peer dependencies to example project
For react-native's native module auotlinking to function properly, all
native dependencies must be direct dependencies of the app. Native
dependencies should never be transitive dependencies.

Also, note that the dependency versions don't match the ones in
packages/zeego/package.json. This is because this example project
doesn't actually depend on the workspace version of zeego, it depends
on 3.0.0-alt.1 which depends on these versions of the native packages.

What really should be done, but which I ran out of time to do, is to
remove the yarn.lock in this example project directory and have this
example app depend on 'workspace:*' version of zeego so that the
monorepo is in a healthy state and this example app runs the code that
is in the monorepo for zeegoo. I attempted to do this and I encountered
build errors compiling expo-modules.

There really should only be 1 yarn.lock in a yarn monorepo.